requirements.txt
# Core data & ML
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # parquet I/O
scikit-learn>=1.3.0

# Forecasting models (Phase 2)
xgboost>=2.0.0
lightgbm>=4.0.0

# Optimization (Phase 4)
ortools>=9.8.3296

# Dashboard (Phase 8)
streamlit>=1.32.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0
click>=8.1.7             # CLI argument parsing
tqdm>=4.66.0             # progress bars
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
